Query pipeline progress and duration directly
This removes the hack-y regular expression for parsing the progress and
duration of the pipeline. I've updated the gst fork to include functions
for querying progress and duration.

Holding off on merging this until I can test it some more. The first few
times I tested this I experienced UI lock-ups and had to kill the
application. Audio still ran so I suspect it was a channel lock up
somewhere. Needs more testing.