Fix off-by-one error in up-next widget
In #13 we fetch the next track fragment when we play the final track
in the current fragment. However, the up-next widget only had room
to display four lines of text. Because there was now four tracks in
the up-next list, the final newline appended to the list hid scrolled
the buffer and hid the first track in the list.

We now trim the final newline from the list before updating the widget.

Fixes #19