Log In with Legacy API and use FFmpeg instead of gstreamer
Pandora appears to have added some new required parameters to the
REST login payload that appear random/encrypted. Fortunately, the
tokens returned by the Legacy JSONv5 API are valid for the REST
API. As a workaround, we use the Legacy JSONv5 API to perform a
Partner Login as an android device and then a User Login to get
a token. Every request after this uses the REST API.

Additionally, swap out gstreamer for FFmpeg+beep. We pipe tracks
to FFmpeg and write them to a temporary file, then use beep to
play the file. github.com/faiface/beep/wav refuses to play  files
with garbage chunk lengths (which FFmpeg sets if you stream to
stdout) so we have to transcode the whole file first.

These two changes are combined because my gstreamer/pkg-config
install appears to be broken so I couldn't test them independently.

Fixes #24
Obsoletes #10, #16, and #23
Partially Addresses #8 and #9